Changed chunk data from floats to uint8 for 4x less data use

+++ extensions/terrain/src/chunk_data.h
@@ -19,8 +19,7 @@ enum class SurfaceState : uint8_t
 
 struct alignas(64) ChunkData
 {
-	// TODO: Change to uint8_t. Use RenderingDevice::DATA_FORMAT_R8_UNORM texture for automatic conversion to float, or raw uint_8 buffer and divide by 255.0
-	std::array<float, terrain_constants::POINTS_VOLUME> points{};
+	std::array<uint8_t, terrain_constants::POINTS_VOLUME> points{};
 	Vector3i position{};
 	SurfaceState surface_state = SurfaceState::EMPTY;
 };
